Extract braze cards into separate component
